chore: separate current components into smaller components
